fix: convert Testimonials to server component to resolve CORS
