+++ src/app/[lang]/page.tsx
@@ -1,7 +1,10 @@
 import Link from "next/link";
 import styles from "./page.module.css";
+import { TestimonialsSection } from "@/components/Testimonials";
+import { Lang } from "need4deed-sdk";
 
-export default function Home() {
+export default async function Home({ params }: { params: Promise<{ lang: Lang }> }) {
+  const lang = (await params).lang;
   return (
     <div className={styles.page}>
       <header>HEADER</header>
@@ -21,6 +24,7 @@ export default function Home() {
           </Link>
         </div>
       </main>
+      <TestimonialsSection lang={lang as Lang} />
       <footer className={styles.footer}>FOOTER</footer>
     </div>
   );
