+++ src/components/Testimonials/Testimonials.tsx
@@ -1,11 +1,8 @@
 import { Lang } from "need4deed-sdk";
-import { useTranslation } from "react-i18next";
-import useTestimonials from "../../hooks/api/useTestimonials";
-
 import { PaginatedCards } from "../core/paginatedCards";
 import TestimonialCard from "./TestimonialCard";
+import fetchTestimonials from "@/hooks/api/testimonials";
 import { ScreenTypes } from "@/config/constants";
-import { useScreenType } from "@/context/DeviceContext";
 
 const cardsPerPageMap = {
   [ScreenTypes.MOBILE]: 1,
@@ -13,10 +10,8 @@ const cardsPerPageMap = {
   [ScreenTypes.DESKTOP]: 3,
 };
 
-export default function Testimonials() {
-  const { i18n } = useTranslation();
-  const [testimonials] = useTestimonials(i18n.language as Lang);
-  const screenType = useScreenType();
+export default async function Testimonials({ lang }: { lang: Lang }) {
+  const testimonials = await fetchTestimonials(lang);
 
   const cards = testimonials.map((t) => <TestimonialCard {...t} key={t.name} />);
 
@@ -26,7 +21,7 @@ export default function Testimonials() {
       arrowButtonColor="orchid-dark"
       bottomIndicatorColor="orchid-dark"
       bottomCurrentIndicatorColor="aubergine-light"
-      cardsPerPage={cardsPerPageMap[screenType]}
+      cardsPerPageMap={cardsPerPageMap}
     />
   );
 }
