+++ src/components/Testimonials/TestimonialsSection.tsx
@@ -1,13 +1,14 @@
+import { Lang } from "need4deed-sdk";
 import { FullWidthContainer, SectionContainer } from "../styled/container";
 import Header from "./Header";
 import Testimonials from "./Testimonials";
 
-export function TestimonialsSection() {
+export function TestimonialsSection({ lang }: { lang: Lang }) {
   return (
     <FullWidthContainer id="testimonials-fullWidthContainer" background-color="var(--color-magnolia)">
       <SectionContainer id="testimonials-container">
         <Header />
-        <Testimonials />
+        <Testimonials lang={lang} />
       </SectionContainer>
     </FullWidthContainer>
   );
